+++ extensions/healthie/actions/checkScheduledAppointments/checkScheduledAppointments.ts
@@ -1,0 +1,77 @@
+import { type Action } from '@awell-health/extensions-core'
+import { Category } from '@awell-health/extensions-core'
+import { HealthieError, mapHealthieToActivityError } from '../../errors'
+import { getSdk } from '../../gql/sdk'
+import { initialiseClient } from '../../graphqlClient'
+import { type settings } from '../../settings'
+import { dataPoints, fields } from './config'
+
+export const checkScheduledAppointments: Action<
+  typeof fields,
+  typeof settings,
+  keyof typeof dataPoints
+> = {
+  key: 'checkScheduledAppointments',
+  category: Category.EHR_INTEGRATIONS,
+  title: 'Check scheduled appointments',
+  description:
+    'Check does the patient have active appointment scheduled for an appointment type',
+  fields,
+  dataPoints,
+  previewable: true,
+  onActivityCreated: async (payload, onComplete, onError): Promise<void> => {
+    const { fields, settings } = payload
+    const { patientId, appointmentTypeId } = fields
+    try {
+      const client = initialiseClient(settings)
+      if (client !== undefined) {
+        const sdk = getSdk(client)
+        const { data } = await sdk.getScheduledAppointments({
+          user_id: patientId,
+          appointment_type_id: appointmentTypeId,
+        })
+        const appointments = data.appointments ?? []
+        const isScheduled = appointments.length > 0
+        await onComplete({
+          data_points: {
+            isScheduled: String(isScheduled),
+          },
+        })
+      } else {
+        await onError({
+          events: [
+            {
+              date: new Date().toISOString(),
+              text: { en: 'API client requires an API url and API key' },
+              error: {
+                category: 'MISSING_SETTINGS',
+                message: 'Missing api url or api key',
+              },
+            },
+          ],
+        })
+      }
+    } catch (err) {
+      if (err instanceof HealthieError) {
+        const errors = mapHealthieToActivityError(err.errors)
+        await onError({
+          events: errors,
+        })
+      } else {
+        const error = err as Error
+        await onError({
+          events: [
+            {
+              date: new Date().toISOString(),
+              text: { en: 'Healthie API reported an error' },
+              error: {
+                category: 'SERVER_ERROR',
+                message: error.message,
+              },
+            },
+          ],
+        })
+      }
+    }
+  },
+}