+++ extensions/healthie/actions/checkScheduledAppointments/config/fields.ts
@@ -1,16 +1,17 @@
 import { type Field, FieldType } from '@awell-health/extensions-core'
 
 export const fields = {
-  id: {
-    id: 'id',
-    label: 'Tag ID',
-    description: 'The ID of the tag in Healthie.',
-    type: FieldType.STRING,
-  },
   patientId: {
     id: 'patientId',
     label: 'Patient ID',
     description: 'The ID of the patient in Healthie.',
     type: FieldType.STRING,
+  },
+  appointmentTypeId: {
+    id: 'appointmentTypeId',
+    label: 'Appointment type ID',
+    description: 'The ID of the appointment type.',
+    type: FieldType.STRING,
+    required: true,
   },
 } satisfies Record<string, Field>